Don't set a default fontSize for links
Setting a default fontSize lead to a weird interface when actually used.
For example, if you stuck a Link inside of a Paragraph with variant
'tiny', the link would be much bigger than the rest of the text. You
could fix this by setting fontSize to inherit on the Link but this felt
like an antipattern.