Clean up how we use theme in tests
For some time we've been passing in the theme object manually when
testing components within Arbor. This was easy enough to get up and
running using emotion-theming but it's become cumbersome. This refactors
and extracts a test helper called `createWithTheme` that can be used to
inject a theme using context via the emotion-theming `ThemeProvider`
component.