Upgrade emotion to emotion 10
I ran into one hiccup while doing this with Tooltips.

Tooltip had to be updated to use `withTheme` in order to pass theme
into the `Text` component. I'm honestly not 100% sure why this is now.
It seems like something is happening between Tippy and emotion where the
theme is not available when the `Text` component is rendered by Tippy. I
tried `react-tippy`'s `useContext` prop but this did not help.

[#162257329]