Update button styles
Our Button with Icon styles were all out of whack. This was to resolve
issues with one of our custom icons being vastly different from the
other icons. This resolves updates the button to work properly with all
icons as opposed to the single case. The icon in question has also been
fixed itself in fort-awesome.

Co-Authored-by: Anthony Costanzo <anthony@catchandrelease.com>