Explicitly pass size prop to Avatar subcomponents
Before, we would indirectly pass size to the `StyledAvatar` and
`StyleAvatarText` by passing along all props we received down to them.
This lead to odd behavior when trying to style the `StyledAvatar` using
styled-system.

Take the following example:

```js
<Avatar mr="regular" name="Ralph Bot" />
```

The margin-right regular would then be applied to the `StyledAvatar`
component as well as the `StyledAvatarText` component which would lead
to unusual results. We only want the styled-system props to apply to the
`StyledAvatar` component.