Fix Avatar borders issues
Order of styled-system style props on the Box component prevented
borderColors from working. Order is IMPORTANT when leveraging
styledSystem style props. Be careful alphabetizing, especially if you
auto alphabetize declarations with a macro... like I did :(

In CSS you can override a generic style by adding a another style below
it in the same declaration. For example:

```
.myBox {
  border: 1px solid #00000;
  border-color: #00FF00;
}
```

will result in a 1px solid border with a color of #00FF00 (nothing new
here).

-------

The problem occurred due to the following declaration in Box

```
const Box = styled.div`
  box-sizing: border-box;
  ${borderColor};
  ${borderRadius};
  ${borders};
  ...
`
```

Note that border color is above borders. If you were to leverage both
the borderColor and border style props on a component they will be
slotted into the output CSS in that order. For Example:

<Box border="1px solid" borderColor="#00ff00" />

will result in:

Box {
  border-color: #00FF00;
  border: 1px solid;
}

"border" is a shorthand declaration, expanding to border width, style,
and color. It defaults color to black and overrides the original
declaration of #00FF00 since it comes AFTER border-color in the style
declaration.