Fix shouldCloseOnOverlayClick Modal bug
The calculated width on the modal was preventing the ability to close
the modal by clicking the area outside of the modal content. This was
because the modal itself was being drawn much wider than the actual
content. So it looked like you could click the overlay even though you
couldn't.

[#167382149]