Translate innerRef to ref in MenuItem and Pane
This cleans up some console errors regarding deprecation of innerRef from
Emotion.  These could be reproduced by opening the Selects and mousing over
items in the AutoComplete stories in Storybook.

This was a little weird to track down because I initially thought to look for
usages of innerRef to _remove_, but what we really want to do is _handle_
innerRef by translating it to ref.

[#170346559]